pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.3",
    "joblib>=1.5.2",
    "loguru>=0.7.3",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "scikit-learn>=1.7.2",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
